wip: enhance JSONL conversion functionality
